Fix context switching.
With this patch context switching finally works.
Without any sense of what's reasonable the OS will alternate between
two tasks: "init", which for now is the same stub printing info about
the kernels status, and "idle", which simply idles and should only be
called when nothing else can be done.
However, as the scheduler simply round robin's all tasks without any
differentiation, this is the state for now.

* kern/main.c: Reduce init task stack size.
* kern/task.h: Tasks have names now.
* kern/task.c: Rewrite cpu mode switching code.
  Never fully replace a PSR, instead only ever switch the cpu mode
  through running `msr` with a corresponding literal.